Tile placements get checked
Now, if a tile gets placed, there will first be checked if the placement is possible. If it is possible the tile will be placed, else nothing changes. Tile placements get logged to the console. There is a bug with placements on the outer tiles, have to find a solution for it.